fix(tabs): using correct style
